requirements.txt
Pillow > 8.3.0
requests >= 2.7.0
